Refactor and pass tests for #6
