+++ test/unit/Helper/DataProvider/StringProvider.php
@@ -1,0 +1,16 @@
+<?php
+namespace Gt\Session\Test\Helper\DataProvider;
+
+trait StringProvider {
+	public function data_randomString():array {
+		$data = [];
+
+		for($i = 0; $i < 10; $i++) {
+			$row = [];
+			$row []= uniqid("random");
+			$data []= $row;
+		}
+
+		return $data;
+	}
+}
