+++ test/unit/SessionTest.php
@@ -3,10 +3,17 @@ namespace Gt\Session\Test;
 
 use Gt\Session\Handler;
 use Gt\Session\Session;
+use Gt\Session\Test\Helper\DataProvider\ConfigProvider;
+use Gt\Session\Test\Helper\DataProvider\StringProvider;
 use Gt\Session\Test\Helper\FunctionMocker;
+use Gt\Session\Test\Helper\DataProvider\KeyValuePairProvider;
 use PHPUnit\Framework\TestCase;
 
 class SessionTest extends TestCase {
+	use KeyValuePairProvider;
+	use StringProvider;
+	use ConfigProvider;
+
 	public function setUp() {
 		FunctionMocker::mock("session_start");
 	}
@@ -53,25 +60,117 @@ class SessionTest extends TestCase {
 		$session->remove("test-key");
 	}
 
-	public function data_randomConfig():array {
-		$data = [];
-		$configKeyList = [
-			"save_path","name","cookie_lifetime","cookie_path",
-			"cookie_domain","cookie_secure","cookie_httponly",
-		];
+	/**
+	 * @dataProvider data_randomString
+	 */
+	public function testGetReturnsNull(string $randomString):void {
+		$handler = $this->getMockBuilder(Handler::class)
+			->getMock();
+		$session = new Session($handler);
+		self::assertNull($session->get($randomString));
+	}
 
-		for($i = 0; $i < 10; $i++) {
-			$row = [];
-			$configItem = [];
+	/**
+	 * @@dataProvider data_randomKeyValuePairs
+	 */
+	public function testSetGet(array $keyValuePairs):void {
+		$handler = $this->getMockBuilder(Handler::class)
+			->getMock();
+		$session = new Session($handler);
 
-			foreach($configKeyList as $configKey) {
-				$configItem[$configKey] = uniqid($configKey);
-			}
-
-			$row []= $configItem;
-			$data []= $row;
+		foreach($keyValuePairs as $key => $value) {
+			$session->set($key, $value);
 		}
 
-		return $data;
+		foreach($keyValuePairs as $key => $value) {
+			self::assertEquals($value, $session->get($key));
+		}
+	}
+
+	/**
+	 * @@dataProvider data_randomKeyValuePairs
+	 */
+	public function testSetGetNotExistsOtherKey(array $keyValuePairs):void {
+		$handler = $this->getMockBuilder(Handler::class)
+			->getMock();
+		$session = new Session($handler);
+
+		foreach($keyValuePairs as $key => $value) {
+			$session->set($key, $value);
+		}
+
+// Test that other keys' values do not match.
+		foreach($keyValuePairs as $key => $value) {
+			do {
+				$otherKey = array_rand($keyValuePairs);
+			}while($otherKey === $key);
+
+			self::assertNotEquals($value, $session->get($otherKey));
+		}
+
+// Test that unknown keys' values do not match.
+		foreach($keyValuePairs as $key => $value) {
+			self::assertNotEquals(
+				$value,
+				$session->get("$key-oh-no")
+			);
+		}
+	}
+
+	/**
+	 * @dataProvider data_randomKeyValuePairs
+	 */
+	public function testContains(array $keyValuePairs):void {
+		$handler = $this->getMockBuilder(Handler::class)
+			->getMock();
+		$session = new Session($handler);
+
+		foreach($keyValuePairs as $key => $value) {
+			$session->set($key, $value);
+		}
+
+		foreach($keyValuePairs as $key => $value) {
+			self::assertTrue($session->contains($key));
+		}
+	}
+
+	/**
+	 * @dataProvider data_randomKeyValuePairs
+	 */
+	public function testNotContains(array $keyValuePairs):void {
+		$handler = $this->getMockBuilder(Handler::class)
+			->getMock();
+		$session = new Session($handler);
+
+		foreach($keyValuePairs as $key => $value) {
+			$session->set($key, $value);
+		}
+
+		foreach($keyValuePairs as $key => $value) {
+			self::assertFalse($session->contains("$key-oh-no"));
+		}
+	}
+
+	/**
+	 * @dataProvider data_randomKeyValuePairs
+	 */
+	public function testDelete(array $keyValuePairs):void {
+		$handler = $this->getMockBuilder(Handler::class)
+			->getMock();
+		$session = new Session($handler);
+
+		foreach($keyValuePairs as $key => $value) {
+			$session->set($key, $value);
+		}
+
+		uasort($keyValuePairs, function () {
+			return rand(-1, 1);
+		});
+
+		foreach($keyValuePairs as $key => $value) {
+			self::assertTrue($session->contains($key));
+			$session->remove($key);
+			self::assertFalse($session->contains($key));
+		}
 	}
 }
